# FinVault/swagger_payment_fixed.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "FinVault — Payment Service",
                 "description":  "Payment Processing, Risk \u0026 Fraud Detection API",
                 "version":  "v1"
             },
    "paths":  {
                  "/api/external-bills/pay":  {
                                                  "post":  {
                                                               "tags":  [
                                                                            "ExternalBills"
                                                                        ],
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/InitiateExternalBillPaymentCommand"
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/InitiateExternalBillPaymentCommand"
                                                                                                                                }
                                                                                                                 },
                                                                                                   "application/*+json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/InitiateExternalBillPaymentCommand"
                                                                                                                                         }
                                                                                                                          }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK"
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/api/external-bills/{id}/verify":  {
                                                          "put":  {
                                                                      "tags":  [
                                                                                   "ExternalBills"
                                                                               ],
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "uuid"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "requestBody":  {
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/VerifyExternalBillPaymentRequest"
                                                                                                                                              }
                                                                                                                               },
                                                                                                          "text/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/VerifyExternalBillPaymentRequest"
                                                                                                                                       }
                                                                                                                        },
                                                                                                          "application/*+json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/VerifyExternalBillPaymentRequest"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                      }
                                                                                      },
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK"
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
                                       "InitiateExternalBillPaymentCommand":  {
                                                                                  "type":  "object",
                                                                                  "properties":  {
                                                                                                     "userId":  {
                                                                                                                    "type":  "string",
                                                                                                                    "format":  "uuid"
                                                                                                                },
                                                                                                     "cardId":  {
                                                                                                                    "type":  "string",
                                                                                                                    "format":  "uuid"
                                                                                                                },
                                                                                                     "email":  {
                                                                                                                   "type":  "string",
                                                                                                                   "nullable":  true
                                                                                                               },
                                                                                                     "billerName":  {
                                                                                                                        "type":  "string",
                                                                                                                        "nullable":  true
                                                                                                                    },
                                                                                                     "billerCategory":  {
                                                                                                                            "type":  "string",
                                                                                                                            "nullable":  true
                                                                                                                        },
                                                                                                     "billNumber":  {
                                                                                                                        "type":  "string",
                                                                                                                        "nullable":  true
                                                                                                                    },
                                                                                                     "amount":  {
                                                                                                                    "type":  "number",
                                                                                                                    "format":  "double"
                                                                                                                }
                                                                                                 },
                                                                                  "additionalProperties":  false
                                                                              },
...
                                       "VerifyExternalBillPaymentRequest":  {
                                                                                "type":  "object",
                                                                                "properties":  {
                                                                                                   "otpCode":  {
                                                                                                                   "type":  "string",
                                                                                                                   "nullable":  true
                                                                                                               }
                                                                                               },
                                                                                "additionalProperties":  false
                                                                            }
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter your JWT token.",
                                                              "scheme":  "Bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ]
}